requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
openai==2.8.1
orjson==3.11.4
outcome==1.3.0.post0
packaging==24.2
pdf2image==1.17.0
pillow==12.0.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.3.2
gunicorn==23.0.0
reportlab==4.4.10
diskcache==5.6.3
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
alembic==1.14.1
redis==5.2.1
Flask-Session==0.8.0
sentry-sdk[flask]==2.19.2
gevent==24.11.1
